requirements.txt
requests==2.22.0
beautifulsoup4==4.9.3
PyYAML==5.3.1
